set page memory address to 64bit

+++ TestDrive/Memory.h
@@ -8,16 +8,16 @@ public:
 	CMemory(void);
 	virtual ~CMemory(void);
 
-	STDMETHOD_(BOOL, Create)(DWORD dwByteSize, LPCTSTR lpszName);
+	STDMETHOD_(BOOL, Create)(uint64_t dwByteSize, LPCTSTR lpszName);
 	STDMETHOD_(void, Release)(void);
 	STDMETHOD_(BOOL, IsInitialized)(void);
-	STDMETHOD_(DWORD, GetSize)(void);
-	STDMETHOD_(BOOL, IsValidAddress)(DWORD dwAddress);
-	STDMETHOD_(BYTE*, GetPointer)(DWORD dwAddress = 0, DWORD dwSize = 0);
-	STDMETHOD_(BOOL, Load)(MEM_DISC disc, LPCTSTR lpszFileName, DWORD offset);
-	STDMETHOD_(BOOL, Save)(MEM_DISC disc, LPCTSTR lpszFileName, DWORD offset, DWORD size, DWORD stride = 0);
-	STDMETHOD_(BOOL, LoadImage)(LPCTSTR lpszFileName, DWORD dwOffset, COLORFORMAT iColorType, DWORD stride = 0);
-	STDMETHOD_(BOOL, StoreImage)(LPCTSTR lpszFileName, DWORD dwOffset,  DWORD dwWidth, DWORD dwHeight, COLORFORMAT iColorType, BOOL bStoreAlpha = FALSE, DWORD stride = 0);
+	STDMETHOD_(uint64_t, GetSize)(void);
+	STDMETHOD_(BOOL, IsValidAddress)(uint64_t dwAddress);
+	STDMETHOD_(BYTE*, GetPointer)(uint64_t dwAddress = 0, uint64_t dwSize = 0);
+	STDMETHOD_(BOOL, Load)(MEM_DISC disc, LPCTSTR lpszFileName, uint64_t offset);
+	STDMETHOD_(BOOL, Save)(MEM_DISC disc, LPCTSTR lpszFileName, uint64_t offset, uint64_t size, DWORD stride = 0);
+	STDMETHOD_(BOOL, LoadImage)(LPCTSTR lpszFileName, uint64_t dwOffset, COLORFORMAT iColorType, DWORD stride = 0);
+	STDMETHOD_(BOOL, StoreImage)(LPCTSTR lpszFileName, uint64_t dwOffset,  DWORD dwWidth, DWORD dwHeight, COLORFORMAT iColorType, BOOL bStoreAlpha = FALSE, DWORD stride = 0);
 	STDMETHOD_(TESTDRIVE_CONFIG*, GetConfig)(void);
 	STDMETHOD_(LPCTSTR, GetName)(void);
 
@@ -27,13 +27,13 @@ public:
 
 protected:
 	BYTE*				m_pMemory;
-	DWORD				m_dwMemorySize;
+	uint64_t				m_dwMemorySize;
 
 private:
 	HANDLE				m_MemHandle;
 	BYTE*				m_pSharedMemory;
 	static CMemory*		m_pHead;
-	CMemory*			m_pNext;
+	CMemory*				m_pNext;
 	CString				m_sName;
 
 	void SetNext(CMemory* pMemory);
